Start Windows fixes I found out that my code was broken on Windows and implemented a few changes for that. I added a manifest which hopefully ensures UTF-8 codepage on newer Windows (which means less hacks for me, on 1903 or later).
I also made it test the codepage, and tell the program if UTF-8 is not active for whatever reason (like the manifest being broken).

This adds buildsystem workarounds to support adding in the manifest (via a resource file).

Any arch-specific files are in the arch/ folder (which currently just has arch/win which contains the manifest and resource file).

I have not implemented a workaround for filename reconversion if not running on a UTF-8 clean version of Windows.

The manifest does exactly 2 things: it requests a UTF-8 codepage on Windows versions that support it, and tells Windows we hypothetically support Vista through 10/11. If we don't do that, Windows 8.1 and newer will lie to us and claim to be 8.0 in the GetVersion() function.