Start implementing argument reconversion Adds a reconv_cli_arg() function implementing preliminary argument reconversion from UTF-16. It just re-returns the original argument if not on Windows.
This implementation is not used yet, I'm too lazy to implement checking for it right now.

I have confirmed that I need 1903 to run fully UTF-8 native (ignoring GDI, which isn't capable of being clean). The C library can run native on 1803 (or supposedly Windows pre-10 if I bundle the C library), but it would give arguments in the ANSI codepage (not UTF-8) if that is true.

I need to implement a check to make sure locale setup works correctly within the code. (That's partially to determine if I would really need to bundle the C libary to work correctly on Windows 7/8.x. I want to be able to save to emoji-named image files or something weird like that.)

P.S. Working correctly on Windows 1507-1709 supposedly requires statically linking the C library from an SDK >= 1803, but I don't really want to do that.