Attempt to have Ctrl-C handling This implements a signal handler for Ctrl-C, which (at least on Linux) allows making sure the file is not created as empty when interrupted by Ctrl-C.
It is not tested on Windows. It will either work, or it will fail to delete due to already being open (I can't tell which). Alternatives are either to expose outfile stream globally and close it explicitly first, or to use unlink() from POSIX (which is called _unlink on Windows), which is nonstandard but behaves the same on all platforms where implemented.